"Displays correctly" test generator working w/ style blacklist
Next I will add an optional argument to change the window size, both in the selenium run when it is generating the test, and in the cypress test itself. Hopefully this way, I can remove pixel-value styles from the style blacklist, as they were previously causing problems from the selenium window and the cypress viewport being different sizes.